Extract verify, discardsession, remakeolm into crypto
